requirements.txt
flask
waitress
nltk
requests
pyyaml